pyproject.toml
[project]
name = "idea-refiner"
version = "0.1.0"
description = "AI-powered micro-SaaS idea generator — multi-model brainstorming pipeline that generates, judges, and refines business ideas until they're worth building"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.17.0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.6",
]

[project.scripts]
idea-refiner = "agents.idea_refiner:main"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
